Rename progressive_tree.rs
