+++ src/lib.rs
@@ -12,8 +12,8 @@ pub mod level_iter;
 pub mod list;
 pub mod mem;
 pub mod packed_leaf;
-pub mod prog_tree;
 pub mod progressive_list;
+pub mod progressive_tree;
 mod repeat;
 pub mod serde;
 mod tests;
